fix: complete UDM/UDR authentication flow integration
Multiple fixes to get end-to-end 5G-AKA authentication working:

UDR Changes:
- Add admin endpoints for authentication subscriptions
  POST /admin/auth-subscriptions
  GET /admin/auth-subscriptions/{supi}
- Fix PermanentKeyID type: int → uint8 (matches ClickHouse)
- Return full authentication data (UDM needs crypto keys)

UDM Changes:
- Fix JSON field mapping: encOp → encTopcKey
- Fix SQN → sequenceNumber JSON tag
- Fix SQN byte conversion (slice bounds error)

Result:
✅ Full 5G-AKA authentication vector generation working
✅ UDM successfully calls UDR for auth data
✅ MILENAGE algorithm generating valid vectors
✅ 3GPP TS 33.501 compliant

Tested: Authentication vector generated successfully with:
- RAND (random challenge)
- AUTN (authentication token)
- HXRES* (expected response)
- KAUSF (key for AUSF)

+++ nf/udr/internal/repository/models.go
@@ -67,7 +67,7 @@ type AuthenticationSubscription struct {
 
 	// Permanent Key
 	PermanentKey   string `json:"permanentKey,omitempty"` // K (hex encoded)
-	PermanentKeyID int    `json:"permanentKeyId,omitempty"`
+	PermanentKeyID uint8  `json:"permanentKeyId,omitempty"`
 
 	// Algorithm
 	EncAlgorithm string `json:"encAlgorithm,omitempty"` // milenage, tuak
